Promote v6 navbar to production
