+++ src/components/NavBar/index.ts
@@ -16,4 +16,3 @@
  */
 
 export * from "./NavBar";
-export * from "./EXV6NavBar";
