+++ src/global_styl/global.styl
@@ -66,21 +66,6 @@ html, body, #main-content, #main-content > div {
     flex: 1 0 auto;
 }
 
-/*
-body.v6 #main-content {
-    display: flex;
-    flex-direction: column;
-
-    > div:first-child {
-        flex: 0 0 auto;
-    }
-
-    > div:nth-child(2) {
-        flex: 1 1 auto;
-    }
-}
-*/
-
 a, .clickable, .fakelink {
     text-decoration: none;
     themed color link-color
