+++ src/ogs.styl
@@ -30,22 +30,20 @@ navbar-height = var(--navbar-height)
 below-navbar = var(--below-navbar)
 
 :root {
-    --navbar-height: 2rem;
-    --below-navbar: 2.3rem;
-}
-body.v6 {
     --navbar-height: 3rem;
     --below-navbar: 3.3rem;
+}
+body.v6 {
+    // experimental
 }
 
 @media (max-height: 500px) {
     :root {
-        --navbar-height: 1.5rem;
-        --below-navbar: 1.8rem;
-    }
-    body.v6 {
         --navbar-height: 2.0rem;
         --below-navbar: 2.3rem;
+    }
+    body.v6 {
+        // experimental
     }
 }
