+++ src/views/HelpFlows/HelpFlows.styl
@@ -38,6 +38,6 @@
     box-shadow: 0 0 2px 2px var(--primary) !important;
 }
 
-#help-for-profile-edit-page, #help-for-profile-edit-page-exv6 {
+#help-for-profile-edit-page {
     max-width: 15rem;
 }
